Version 2.5.0 Added enable to the schema for the configuration Modified the module loading code and configuration code to support enable as Enabled in the model Enabled = false keeps the image and consequently the window from being rendered. Renamed MFDWindow to AuxWindow and got rid of the most convenient MFD references and replaced MFD with Window in logging.
