--- MFDisplay/MFDWindow.xaml
@@ -1,12 +1,0 @@
-﻿<Window x:Class="MFDisplay.MFDWindow"
-        xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-        xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-        xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-        xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-        xmlns:local="clr-namespace:MFDisplay"
-        mc:Ignorable="d"
-        Title="MFD" Name="mfdWindow" BorderThickness="0" Topmost="True" Loaded="Window_Loaded" Closed="Window_Closed" ResizeMode="NoResize" WindowStyle="None" ShowInTaskbar="False" Opacity="0" AllowsTransparency="True">
-    <Grid>
-        <Image Width="auto" Height="auto" x:Name="imgMain" Stretch="Fill"  Margin="0"/>
-    </Grid>
-</Window>
