+++ MFDisplay/mfdsettings.config
@@ -16,7 +16,7 @@
   <MFDSettings xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
   xsi:schemaLocation="http://tempuri.org/mfdsettings.xsd mfdsettings.xsd"
   saveClips="False" filePath="E:\HOTAS\TARGET\CTS\Docs\Profile JPGs"
-  defaultConfig="F-14P">
+  defaultConfig="BMSFalconF-16">
   
   <DefaultConfigurations>
     <add name="LMFD" opacity="1" width="870" height="700" left="2575"
@@ -394,6 +394,14 @@
     <add moduleName="UH-1HV" displayName="UH-1H Huey HiViz" filename="_High Contrast MFDs\DCS UH1 MFD HiVis.jpg">
       <Configurations>
         <add name="WHKEY" filename="UH-1\DCS UH1 WH.jpg" />
+      </Configurations>
+    </add>
+    <add moduleName="BMSFalconF-16" displayName="BMS Falcon F-16" filename="BMSFalcon-F-16\FalconBMS_Warthog_Throttle_v2.png">
+      <Configurations>
+        <add name="LMFD" enable="false" />
+        <add name="RMFD" enable="false" />
+        <add name="WHKEY" opacity=".8" width="885" height="580" left="5120" top="360" xOffsetStart="0" xOffsetFinish="1079" yOffsetStart="0" yOffsetFinish="764" filename="BMSFalcon-F-16\FalconBMS_Warthog_Throttle_v2.png" />
+        <add name="WHJOYSTICK" opacity=".8" width="885" height="580" left="5120" top="900" xOffsetStart="0" xOffsetFinish="1051" yOffsetStart="0" yOffsetFinish="743" filename="BMSFalcon-F-16\FalconBMS_Warthog_Joystick_V2.png" />
       </Configurations>
     </add>
   </Modules>
